Split schema store into two stores
